// src/types/Match.ts
export const matchInfoSchema = z.object({
	id: z.number(),
	start_time: z.coerce.date(),
	end_time: z.coerce.date().nullable(),
	name: z.string(),
});
// ...
export const userSchema = z.object({
	avatar_url: z.string(),
	country_code: z.string(),
	default_group: z.string(),
	id: z.number(),
	is_active: z.boolean(),
	is_bot: z.boolean(),
	is_deleted: z.boolean(),
	is_online: z.boolean(),
	is_supporter: z.boolean(),
	last_visit: z.string().nullable(),
	pm_friends_only: z.boolean(),
	profile_colour: z.any(),
	username: z.string(),
	country: countrySchema,
});

export const beatmapSchema = z.object({
	beatmapset_id: z.number(),
	difficulty_rating: z.number(),
	id: z.number(),
	mode: z.string(),
	status: z.string(),
	total_length: z.number(),
	user_id: z.number(),
	version: z.string(),
	beatmapset: beatmapsetSchema,
});

export const gameSchema = z.object({
	beatmap_id: z.number(),
	id: z.number(),
	start_time: z.string(),
	end_time: z.string(),
	mode: z.string(),
	mode_int: z.number(),
	scoring_type: z.string(),
	team_type: z.string(),
	mods: z.array(z.string()),
	scores: z.array(scoresSchema),
});

export const resultEventSchema = z.object({
	detail: z.object({
		type: z.literal("other"),
		text: z.string().nullable(),
	}),
	id: z.number(),
	timestamp: z.coerce.date(),
	user_id: z.null(),
	game: gameSchema,
});

export type ResultEvent = z.infer<typeof resultEventSchema>;

export const disbandedEventSchema = z.object({
	detail: z.object({
		type: z.literal("match-disbanded"),
	}),
	id: z.number(),
	timestamp: z.coerce.date(),
	user_id: z.null(),
});

export type DisbandedEvent = z.infer<typeof disbandedEventSchema>;

export const eventSchema = z.union([resultEventSchema, disbandedEventSchema]);

export type Event = z.infer<typeof eventSchema>;

export const matchSchema = z.object({
	match: matchInfoSchema,
	events: z.array(eventSchema),
	users: z.array(userSchema),
	first_event_id: z.number(),
	latest_event_id: z.number(),
	current_game_id: z.any(),
});

export type OsuUser = z.infer<typeof userSchema>;
export type MatchData = z.infer<typeof matchSchema>;

export const rawMatchSchema = matchSchema.extend({
	events: z
		.object({
			id: z.number(),
		})
		.passthrough()
		.array(),
});